requirements.txt
openai
python-dotenv
customtkinter
pygame
pillow
darkdetect 